+++ src/Organizations/constants.js
@@ -1,8 +1,8 @@
-import React from 'react';
 import { FormattedMessage } from 'react-intl';
 
 export const ORGANIZATION_SECTIONS = {
   summarySection: 'summarySection',
+  bankingInformationSection: 'bankingInformationSection',
   contactInformationSection: 'contactInformationSection',
   contactPeopleSection: 'contactPeopleSection',
   interfacesSection: 'interfacesSection',
@@ -12,10 +12,12 @@ export const ORGANIZATION_SECTIONS = {
   notesSection: 'notesSection',
   integrationDetailsSection: 'integrationDetailsSection',
   agreements: 'linkedAgreements',
+  privilegedDonorInformation: 'privilegedDonorInformation',
 };
 
 export const ORGANIZATION_SECTION_LABELS = {
   [ORGANIZATION_SECTIONS.summarySection]: <FormattedMessage id="ui-organizations.summary" />,
+  [ORGANIZATION_SECTIONS.bankingInformationSection]: <FormattedMessage id="ui-organizations.bankingInformation" />,
   [ORGANIZATION_SECTIONS.contactInformationSection]: <FormattedMessage id="ui-organizations.contactInformation" />,
   [ORGANIZATION_SECTIONS.contactPeopleSection]: <FormattedMessage id="ui-organizations.contactPeople" />,
   [ORGANIZATION_SECTIONS.interfacesSection]: <FormattedMessage id="ui-organizations.interface" />,
@@ -24,6 +26,7 @@ export const ORGANIZATION_SECTION_LABELS = {
   [ORGANIZATION_SECTIONS.accountsSection]: <FormattedMessage id="ui-organizations.accounts" />,
   [ORGANIZATION_SECTIONS.integrationDetailsSection]: <FormattedMessage id="ui-organizations.integrationDetails" />,
   [ORGANIZATION_SECTIONS.agreements]: <FormattedMessage id="ui-organizations.linkedAgreements.section" />,
+  [ORGANIZATION_SECTIONS.privilegedDonorInformation]: <FormattedMessage id="ui-organizations.privilegedDonorInformation" />,
 };
 
 export const CREATE_UNITS_PERM = 'organizations.acquisitions-units-assignments.assign';
@@ -41,3 +44,5 @@ export const MAP_FIELD_ACCORDION = {
   agreements: ORGANIZATION_SECTIONS.vendorTermsSection,
   accounts: ORGANIZATION_SECTIONS.accountsSection,
 };
+
+export const BANKING_INFORMATION_FIELD_NAME = 'bankingInformation';