+++ src/common/hooks/useEventEmitter/useEventEmitter.js
@@ -1,0 +1,7 @@
+import { EventEmitter } from '../../utils';
+
+const eventEmitter = new EventEmitter();
+
+export const useEventEmitter = () => {
+  return eventEmitter;
+};